requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
tenacity==9.1.2
langdetect==1.0.9
requests==2.32.4
pycryptodome==3.23.0
psutil==7.1.0
sqlalchemy==2.0.43
pyyaml==6.0.2
pytest==8.4.1
httpx==0.28.1
Pillow==11.3.0
pytesseract==0.3.13

